% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{anndata_check_load}
\alias{anndata_check_load}
\title{Checks if anndata package is loaded}
\usage{
anndata_check_load()
}
\description{
If called and python environment is not set up, this is realized. Else, it checks if the anndata
package is loaded, and if not, it does this.
}
