% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{anndata_is_identical}
\alias{anndata_is_identical}
\title{Check if two anndata objects are identical}
\usage{
anndata_is_identical(a, b)
}
\arguments{
\item{a}{Anndata object one}

\item{b}{Anndata object two}
}
\value{
boolean whether they are identical
}
\description{
Check if two anndata objects are identical
}
