% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{AnnData_to_SingleCellExperiment}
\alias{AnnData_to_SingleCellExperiment}
\title{Convert AnnData to SingleCellExperiment}
\usage{
AnnData_to_SingleCellExperiment(ad)
}
\arguments{
\item{ad}{AnnData object}
}
\value{
SingleCellObject
}
\description{
Convert AnnData to SingleCellExperiment
}
