% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneS.R
\name{autogenes_checkload}
\alias{autogenes_checkload}
\title{Checks if python and the autogenes module are available and installs them if they are not.}
\usage{
autogenes_checkload()
}
\description{
Checks if python and the autogenes module are available and installs them if they are not.
}
