% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batch_ids}
\alias{batch_ids}
\title{Batch ids for the example single-cell RNA-seq dataset from Maynard et al.}
\format{
vector with 300 elements
}
\source{
Maynard et al. (2020), Cell,
https://www.sciencedirect.com/science/article/pii/S0092867420308825
}
\usage{
batch_ids
}
\description{
The batch ids (in this case sample ids) are in the same order as cells in the sc RNA-seq matrix
}
\keyword{datasets}
