% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batch_ids_2}
\alias{batch_ids_2}
\title{Patient ids Number 2 from Hao et al.}
\format{
A vector with the patient ids. In the same order as the columns
}
\source{
Hao et al. (2021): Integrated analysis of multimodal single-cell data.
In Cell 184 (13), 3573-3587.e29. DOI: 10.1016/j.cell.2021.04.048.
}
\usage{
batch_ids_2
}
\description{
The identifiers of patients of whom the cells were taken
}
\keyword{datasets}
