% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisque.R
\name{bisque_reference_decomp}
\alias{bisque_reference_decomp}
\title{Calculates the decomposition using the bisque algorithm}
\usage{
bisque_reference_decomp(
  bulk_eset,
  signature_matrix,
  single_cell_object = NULL,
  cell_type_annotations = NULL,
  verbose = T
)
}
\arguments{
\item{bulk_eset}{An Expression Set containing bulk data.}

\item{signature_matrix}{The Signature matrix.}

\item{single_cell_object}{A Matrix with the single-cell data. Rows are genes and columns are samples.}

\item{cell_type_annotations}{A Vector of the cell type annotations. Has to be in the same order as the samples in single_cell_object.}

\item{verbose}{Whether the algorithm should print out what it is doing.}
}
\value{
A list. Slot bulk.props contains a matrix of cell type proportion estimates with cell types as rows and individuals as columns. Slot sc.props contains a matrix of cell type proportions estimated directly from counting single-cell data. Slot rnorm contains Euclidean norm of the residuals for each individual’s proportion estimates. Slot genes.used contains vector of genes used in decomposition.Slot transformed.bulk contains the transformed bulk expression used for decomposition. These values are generated by applying a linear transformation to the CPM expression.
}
\description{
Generates a reference profile based on single-cell data. Learns a transformation of bulk expression based  on  observed  single-cell  proportions  and  performs  NNLS  regression  on  these  transformed values to estimate cell proportions.
}
