% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSeq_sc.R
\name{bseqsc_config}
\alias{bseqsc_config}
\title{Setup BSeq-SC External Dependencies}
\usage{
bseqsc_config(file = NULL, error = FALSE)
}
\arguments{
\item{file}{Path to the CIBERSORT source R code.}

\item{error}{Logical that indicates if an error should be thrown
if configuration failed.}
}
\value{
The path where the file was copied, or \code{NULL} if \code{bseqsc} is not correctly
configured.
}
\description{
Configure \code{bseqsc} by setting up \code{CIBERSORT} source code.
}
\details{
\code{BSeq-sc} uses \code{CIBERSORT} to estimate cell type proportions, based on reference
expression profiles.
Due to licensing requirements, source code for this algorithm needs to be
downloaded separately from its website http://cibersort.stanford.edu.
It is released under the Stanford Non-Commercial License.
In order to use it with \code{bseqsc} you will need to:
\enumerate{
\item Got to http://cibersort.stanford.edu
\item Register and log in
\item Download the latest R source code from the \href{http://cibersort.stanford.edu/download.php}{Download section}.
\item Configure \code{bseqsc} by pointing it to the downloaded file. This is
done using the function \code{bseqsc_config}, which will copy the given R
source file into the \code{R-data/bseqsc} sub-directory in the user's home
directory for subsequent usage:
}\preformatted{bseqsc_config('path/to/downloaded/source/CIBERSORT.R')
}
}
