% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{build_anndata}
\alias{build_anndata}
\title{Create anndata object from single cell expression matrix, cell type labels and gene symbols}
\usage{
build_anndata(x, obs, var, obsm = NULL, varm = NULL)
}
\arguments{
\item{x}{single cell expression matrix, rows = genes, columns = samples}

\item{obs}{vector of cell type labels}

\item{var}{vector of gene symbols}

\item{obsm}{Key-indexed multi-dimensional observations annotation of length #observations.}

\item{varm}{Key-indexed multi-dimensional variables annotation of length #variables.}
}
\value{
AnnData object
}
\description{
Create anndata object from single cell expression matrix, cell type labels and gene symbols
}
