% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\name{build_model}
\alias{build_model}
\title{Building the signature matrix}
\usage{
build_model(
  single_cell_object,
  cell_type_annotations = NULL,
  batch_ids = NULL,
  method = deconvolution_methods,
  bulk_gene_expression = NULL,
  verbose = TRUE,
  cell_type_column_name = NULL,
  markers = NULL,
  ...
)
}
\arguments{
\item{single_cell_object}{A matrix or dataframe with the single-cell data. Rows are genes,
columns are samples. Row and column names need to be set. Alternatively a SingleCellExperiment
or an AnnData object can be provided. In that case, note that cell-type labels need to be
indicated either directly providing a vector (cell_type_annotations) or by indicating the
column name that indicates the cell-type labels (cell_type_column_name). (Anndata: obs object,
SingleCellExperiment: colData object)}

\item{cell_type_annotations}{A Vector of the cell type annotations. Has to be in the same order
as the samples in single_cell_object}

\item{batch_ids}{A vector of the ids of the samples or individuals.}

\item{method}{A string specifying the method.
Supported methods are 'bisque', 'momf', 'dwls', 'scaden', 'cibersortx' and 'autogenes'}

\item{bulk_gene_expression}{A matrix of bulk data. Rows are genes, columns are samples. Necessary
for MOMF, defaults to NULL. Row and column names need to be set}

\item{verbose}{Whether the algorithms should print out what they are doing}

\item{cell_type_column_name}{Name of the column in (Anndata: obs, SingleCellExperiment: colData),
that contains the cell-type labels. Is only used if no cell_type_annotations vector is
provided.}

\item{markers}{Named list of cell type marker genes.
The type of gene identifiers (names(markers)) must be the same as the ones used as feature/row names in the
single_cell_object.}

\item{...}{Additional parameters, passed to the algorithm used}
}
\value{
The signature matrix. Rows are genes, columns are cell types
}
\description{
The single_cell_object is expected to have rownames() and colnames()
}
