% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneS.R
\name{build_model_autogenes}
\alias{build_model_autogenes}
\title{Calculates the signature model with AutoGeneS}
\usage{
build_model_autogenes(
  single_cell_object,
  cell_type_annotations,
  bulk_gene_expression = NULL
)
}
\arguments{
\item{single_cell_object}{A matrix or dataframe with the single-cell data. Rows are genes, columns are samples. Row and column names need to be set.}

\item{cell_type_annotations}{A Vector of the cell type annotations. Has to be in the same order as the samples in single_cell_object}

\item{bulk_gene_expression}{A matrix of bulk data. Rows are genes, columns are samples. Necessary for MOMF, defaults to NULL.Row and column names need to be set.}
}
\value{
The signature matrix. Rows are genes, columns are cell types.
}
\description{
Calculates the signature model with AutoGeneS
}
