% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPrism.R
\name{build_model_bayesprism}
\alias{build_model_bayesprism}
\title{No model is built as BayesPrism performs signature building and deconvolution in one step.}
\usage{
build_model_bayesprism()
}
\value{
NULL.
}
\description{
Please use the deconvolute method with your single cell and bulk rna seq data to use BayesPrism
}
