% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisque.R
\name{build_model_bisque}
\alias{build_model_bisque}
\title{No model is build as Bisque does both steps in one.}
\usage{
build_model_bisque()
}
\value{
NULL.
}
\description{
Please use the deconvolute method with your single cell and bulk rna seq data to use Bisque
}
