% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSEQ_sc.R
\name{build_model_bseqsc}
\alias{build_model_bseqsc}
\title{Calculates the signature model with BSeq-sc}
\usage{
build_model_bseqsc(
  single_cell_object,
  cell_type_annotations,
  markers,
  batch_ids,
  ct_scale = TRUE,
  limit = TRUE
)
}
\arguments{
\item{single_cell_object}{A matrix with the single-cell data. Rows are genes, columns are
samples. Row and column names need to be set.}

\item{cell_type_annotations}{A vector of the cell type annotations. Has to be in the same order
as the samples in single_cell_object.}

\item{markers}{Named list of cell type marker genes.
The type of gene identifiers (names(markers)) must be the same as the ones used as feature/row
names in the single_cell_object.}

\item{batch_ids}{A vector of the ids of the samples or individuals.}

\item{ct_scale}{logical that indicates if the single cell expresson profiles
should be rescaled according to their cell type average total count.}

\item{limit}{logical that indicates if the returned basis matrix should only contain
cell types listed in \code{markers}.}
}
\value{
The signature matrix. Rows are genes, columns are cell types.
}
\description{
Calculates the signature model with BSeq-sc
}
