% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDSeq.R
\name{build_model_cdseq}
\alias{build_model_cdseq}
\title{No model is build as CDSeq does both steps in one.}
\usage{
build_model_cdseq()
}
\description{
Please use the deconvolute method with your single cell and bulk rna seq data to use CDSeq.
}
