% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CibersortX.R
\name{build_model_cibersortx}
\alias{build_model_cibersortx}
\title{Signature matrix creation with CIBERSORTx}
\usage{
build_model_cibersortx(
  single_cell_object,
  cell_type_annotations,
  container = c("docker", "singularity"),
  container_path = NULL,
  verbose = FALSE,
  input_dir = NULL,
  output_dir = NULL,
  display_heatmap = FALSE,
  k_max = 999,
  ...
)
}
\arguments{
\item{single_cell_object}{A matrix with the single-cell data. Rows are genes, columns are
samples. Row and column names need to be set.}

\item{cell_type_annotations}{A vector of the cell type annotations. Has to be in the same order
as the samples in single_cell_object.}

\item{container}{The container used ot run the method. The possibilities are 'docker' (default) or
'singularity'}

\item{container_path}{the path where the singularity .sif file is/will be stored (optional)}

\item{verbose}{Whether to produce an output on the console.}

\item{input_dir}{The directory in which the input files can be found (or are created in).
Default is a temporary directory.}

\item{output_dir}{The directory in which the output files are saved. Default is a temporary
directory.}

\item{display_heatmap}{Whether to display the generated heatmap.}

\item{k_max}{Maximum condition number (default: 999). Will be added to the file name.}

\item{...}{Additional parameters supplied to the algorithm. Options are:
\itemize{
\item g_min \if{html}{\out{<int>}} Minimum number of genes per cell type in sig. matrix (default: 300).
\item g_max \if{html}{\out{<int>}} Maximum number of genes per cell type in sig. matrix (default: 500).
\item q_value \if{html}{\out{<int>}} Q-value threshold for differential expression (default: 0.01).
\item filter \if{html}{\out{<bool>}} Remove non-hematopoietic genes (default: FALSE).
\item remake \if{html}{\out{<bool>}} Remake signature gene matrix (default: False).
\item replicates \if{html}{\out{<int>}} Number of replicates to use for building scRNAseq reference file
(default: 5).
\item sampling \if{html}{\out{<float>}} Fraction of available single cell GEPs selected using random sampling
(default: 0.5).
\item fraction \if{html}{\out{<float>}} Fraction of cells of same identity showing evidence of expression
(default: 0.75).
}}
}
\value{
The signature matrix. Rows are genes, columns are cell types.
}
\description{
Signature matrix creation with CIBERSORTx
}
