% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM.R
\name{build_model_cpm}
\alias{build_model_cpm}
\title{No model is build as CPM does not use a signature matrix.}
\usage{
build_model_cpm()
}
\value{
NULL.
}
\description{
Please use the deconvolute method with you single cell and bulk rna seq data to use CPM.
}
