% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuSiC.R
\name{build_model_music}
\alias{build_model_music}
\title{No model is build as MuSiC does both steps in one.}
\usage{
build_model_music()
}
\description{
Please use the deconvolute method with you single cell and bulk rna seq data to use MuSiC.
}
