% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{build_model_scaden}
\alias{build_model_scaden}
\title{Builds Scaden Model from scRNA data}
\usage{
build_model_scaden(
  single_cell_object,
  cell_type_annotations,
  bulk_gene_expression,
  model_path = NULL,
  batch_size = 128,
  learning_rate = 1e-04,
  steps = 5000,
  var_cutoff = NULL,
  cells = 100,
  samples = 1000,
  dataset_name = "scaden",
  verbose = FALSE
)
}
\arguments{
\item{single_cell_object}{A matrix with the single-cell data. Rows are genes, columns are
samples. Row and column names need to be set.}

\item{cell_type_annotations}{A vector of the cell type annotations. Has to be in the same order
as the samples in single_cell_object.}

\item{bulk_gene_expression}{A matrix of bulk data. Rows are genes, columns are samples.
Row and column names need to be set.}

\item{model_path}{Path where model directory should be created (optional).}

\item{batch_size}{Training  of model: Batch size to use for training (default: 128).}

\item{learning_rate}{Training of model: Learning rate used for training (default: 1E-4).}

\item{steps}{Training of model: Number of training steps (default: 5000).}

\item{var_cutoff}{Training data processing: Filter out genes with a variance less than the
specified cutoff. A low cutoff is recommended,this should only remove genes that are obviously
uninformative. (default NULL).}

\item{cells}{Bulk simulation: Number of cells per sample (default: 100).}

\item{samples}{Bulk simulation: Number of samples to simulate (default: 1000).}

\item{dataset_name}{Bulk simulation: Name of simulated dataset (default scaden).}

\item{verbose}{Whether to produce an output on the console (default: false).}
}
\value{
The path to the scaden model.
}
\description{
The model is saved in a defined directory.
}
