% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{build_signature_matrix_using_Seurat}
\alias{build_signature_matrix_using_Seurat}
\title{Building the signature matrix using Seurat}
\usage{
build_signature_matrix_using_Seurat(
  scdata,
  id,
  path = NULL,
  verbose = FALSE,
  diff_cutoff = 0.5,
  pval_cutoff = 0.01
)
}
\arguments{
\item{scdata}{The single cell data matrix}

\item{id}{A Vector of the cell type annotations}

\item{path}{OPTIONAL path for saving generated files}

\item{verbose}{Whether to output what DWLS is doing}

\item{diff_cutoff}{The FC cutoff}

\item{pval_cutoff}{The pValue cutoff}
}
\value{
The computed signature matrix
}
\description{
When path = NULL, the generated files in the processes will not be saved and output.
}
