% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk}
\alias{bulk}
\title{Example RNA-seq dataset from Hoek et al.}
\format{
matrix with genes x samples
}
\source{
Hoek et al. (2015), PLOS ONE, https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0118528
}
\usage{
bulk
}
\description{
A dataset of eight RNA-seq samples from patients vaccinated with trivalent inactivated influenza vaccine (TIV)
}
\keyword{datasets}
