% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_small}
\alias{bulk_small}
\title{Small example RNA-seq dataset from Hoek et al.}
\format{
matrix with genes x samples
}
\source{
Hoek et al. (2015), PLOS ONE, https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0118528
}
\usage{
bulk_small
}
\description{
A dataset of eight RNA-seq samples from patients vaccinated with trivalent inactivated influenza vaccine (TIV) including only 100 genes
}
\keyword{datasets}
