% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\name{calc_condition_number}
\alias{calc_condition_number}
\title{Calculation of the condition number}
\usage{
calc_condition_number(signature_matrix)
}
\arguments{
\item{signature_matrix}{A signature matrix created with the build_model method}
}
\value{
The condition number
}
\description{
A problem with a low condition number is said to be well-conditioned, while a problem with a
high condition number is said to be ill-conditioned. An ill-conditioned problem is one where, for
a small change in the inputs (the independent variables) there is a large change in the answer or
dependent variable.
}
\examples{
data("single_cell_data_1")
data("cell_type_annotations_1")
data("batch_ids_1")
data("bulk")

single_cell_data <- single_cell_data_1[1:2000,1:500]
cell_type_annotations <- cell_type_annotations_1[1:500]
batch_ids <- batch_ids_1[1:500]
bulk <- bulk[1:2000,]

signature_matrix_bisque <- build_model(
  single_cell_data, cell_type_annotations, "bisque",
  batch_ids
)
cond_num <- calc_condition_number(signature_matrix_bisque)
cond_num
}
