% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM.R
\name{calculate_cell_embedding}
\alias{calculate_cell_embedding}
\title{Calculation of the cell_space parameter needed by CPM}
\usage{
calculate_cell_embedding(
  single_cell_object,
  cell_type_annotations,
  method = c("PCA", "UMAP", "TSNE")
)
}
\arguments{
\item{single_cell_object}{A matrix with the single-cell data. Rows are genes, columns are
samples. Row and column names need to be set.}

\item{cell_type_annotations}{A vector of the cell type annotations. Has to be in the same order
as the samples in single_cell_object.}

\item{method}{Either "PCA", "UMAP" or "TSNE"}
}
\value{
A matrix with two dimensions. The rows are the cells, the columns are the two dimensions
calculated by Seurat
}
\description{
Calculation of the cell_space parameter needed by CPM
}
