% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_annotations_2}
\alias{cell_type_annotations_2}
\title{Cell Type Annotations Number 2 from Hao et al.}
\format{
A vector with the annotations. In the same order as the columns
}
\source{
Hao et al. (2021): Integrated analysis of multimodal single-cell data.
In Cell 184 (13), 3573-3587.e29. DOI: 10.1016/j.cell.2021.04.048.
}
\usage{
cell_type_annotations_2
}
\description{
The corresponding cell type annotations containing the classes
"B", "CD4 T", "CD8 T", "DC", "Mono" and "NK"
}
\keyword{datasets}
