% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_annotations_small}
\alias{cell_type_annotations_small}
\title{Cell type annotations for the small example sigle-cell RNA-seq dataset from Maynard et al.}
\format{
vector with 50 elements
}
\source{
Maynard et al. (2020), Cell, https://www.sciencedirect.com/science/article/pii/S0092867420308825
}
\usage{
cell_type_annotations_small
}
\description{
cell types are in the same order as cells in the small sc RNA-seq matrix
}
\keyword{datasets}
