% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\name{check_and_install}
\alias{check_and_install}
\title{Checking and installing all dependencies for the specific methods}
\usage{
check_and_install(method)
}
\arguments{
\item{method}{The name of the method that is used}
}
\description{
Checking and installing all dependencies for the specific methods
}
