% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{convert_to_matrix}
\alias{convert_to_matrix}
\title{Converts the object into a matrix}
\usage{
convert_to_matrix(object, cell_type_annotations, cell_type_column_name = NULL)
}
\arguments{
\item{object}{An input matrix, data frame, expression set, etc.}

\item{cell_type_annotations}{A vector of the cell type annotations. Has to be in the same order
as the samples in object. If not used (for example for bulk data), just supply anything, like
a string.}

\item{cell_type_column_name}{Name of the column in (Anndata: obs, SingleCellExperiment: colData),
that contains the cell-type labels. Is only used if no cell_type_annotations vector is
provided.}
}
\value{
The same object, but of type matrix
}
\description{
Converts the object into a matrix
}
