% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CibersortX.R
\name{create_container_command}
\alias{create_container_command}
\title{Creation of the docker command}
\usage{
create_container_command(
  in_dir,
  out_dir,
  container = c("docker", "singularity"),
  sing_container_path = NULL,
  method = c("create_sig", "impute_cell_fractions"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{in_dir}{The folder in which the input files can be found (or are created in). Default is
a temporary directory.}

\item{out_dir}{The directory in which the output files are saved. Default is a temporary
directory.}

\item{container}{The container to use. Possibilities are 'docker' and 'singularity'}

\item{sing_container_path}{The path where the singularity container is stored (optional)}

\item{method}{Which docker command should be be created. For signature matrix creation use
"create_sig", for cell type deconvolution use "impute_cell_fractions".}

\item{verbose}{Whether to produce an output on the console.}

\item{...}{The additional parameters for the command.}
}
\value{
A valid docker command to be run.
}
\description{
Creation of the docker command
}
