% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneS.R
\name{create_tuple_string}
\alias{create_tuple_string}
\title{Creates a string that is evaluated by python as a collection of values}
\usage{
create_tuple_string(vector_of_elements, strings = FALSE)
}
\arguments{
\item{vector_of_elements}{A vector of strings or numbers}

\item{strings}{Whether the input consists of strings that need to be quoted}
}
\value{
The formatted string
}
\description{
Creates a string that is evaluated by python as a collection of values
}
