% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_config.R
\name{create_virtualenv}
\alias{create_virtualenv}
\title{Creates a new virtual environment.}
\usage{
create_virtualenv(path_to_python_binaries)
}
\arguments{
\item{path_to_python_binaries}{Path to the Python binaries.}
}
\value{

}
\description{
Creates a virtual environment and activates it.
The pip version is automatically upgraded to the newest version.
Python3 required for scaden and scanpy to work.
}
