% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{de_analysis_mast}
\alias{de_analysis_mast}
\title{Performing DE analysis using mast}
\usage{
de_analysis_mast(scdata, id, path, verbose = FALSE)
}
\arguments{
\item{scdata}{The single cell data matrix}

\item{id}{A Vector of the cell type annotations}

\item{path}{OPTIONAL path for saving generated files}

\item{verbose}{Whether to print output}
}
\value{
A list with the cell types and their differentially expressed genes
}
\description{
When path = NULL, the generated files in the processes will not be saved and output.
}
