% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{DEAnalysis}
\alias{DEAnalysis}
\title{Performing DE analysis using Seurat}
\usage{
DEAnalysis(scdata, id, path = NULL)
}
\arguments{
\item{scdata}{The single cell data matrix}

\item{id}{A Vector of the cell type annotations}

\item{path}{OPTIONAL path for saving generated files}
}
\value{
List with the differentially expressed genes for each cell type
}
\description{
When path = NULL, the generated files in the processes will not be saved and output
}
