% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisque.R
\name{deconvolute_bisque}
\alias{deconvolute_bisque}
\title{Calculates the decomposition using the bisque algorithm}
\usage{
deconvolute_bisque(
  bulk_gene_expression,
  signature_matrix,
  single_cell_object,
  cell_type_annotations,
  batch_ids,
  markers = NULL,
  cell_types = "cellType",
  subject_names = "batchId",
  use_overlap = FALSE,
  verbose = FALSE,
  old_cpm = TRUE
)
}
\arguments{
\item{bulk_gene_expression}{A matrix or dataframe with the bulk data. Rows
are genes, columns are samples.}

\item{signature_matrix}{The Signature matrix.}

\item{single_cell_object}{A Matrix with the single-cell data. Rows are genes
and columns are samples.}

\item{cell_type_annotations}{A Vector of the cell type annotations. Has to
be in the same order as the samples in single_cell_object.}

\item{batch_ids}{A vector of the ids of the samples or individuals.}

\item{markers}{Structure, such as character vector, containing marker genes
to be used in decomposition. \code{base::unique(base::unlist(markers))} should
return a simple vector containing each gene name. If no argument or NULL
provided, the method will use all available genes for decomposition.}

\item{cell_types}{Character string. Name of phenoData attribute in sc.eset
indicating cell type label for each cell}

\item{subject_names}{Character string. Name of phenoData attribute in sc.eset
indicating individual label for each cell}

\item{use_overlap}{Boolean. Whether to use and expect overlapping samples
in decomposition.}

\item{verbose}{Whether the algorithm should print out what it is doing.}

\item{old_cpm}{Prior to version 1.0.4 (updated in July 2020), the package
converted counts to CPM after subsetting the marker genes. Github user
randel pointed out that the order of these operations should be switched.
Thanks randel! This option is provided for replication of older BisqueRNA
but should be enabled, especially for small marker gene sets.
We briefly tested this change on the cortex and adipose datasets.
The original and new order of operations produce estimates that have an
average correlation of 0.87 for the cortex and 0.84 for the adipose within
each cell type.}
}
\value{
A list. Slot bulk_props contains a matrix of cell type proportion
estimates with cell types as rows and individuals as columns. Slot sc_props
contains a matrix of cell type proportions estimated directly from counting
single-cell data. Slot rnorm contains Euclidean norm of the residuals for
each individual's proportion estimates. Slot genes_used contains vector of
genes used in decomposition.Slot transformed_bulk contains the transformed
bulk expression used for decomposition. These values are generated by
applying a linear transformation to the CPM expression.
}
\description{
Generates a reference profile based on single-cell data. Learns a
transformation of bulk expression based  on  observed  single-cell
proportions  and  performs  NNLS  regression  on  these  transformed
values to estimate cell proportions.
}
