% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSEQ_sc.R
\name{deconvolute_bseqsc}
\alias{deconvolute_bseqsc}
\title{Deconvolution Analysis using BSEQ-sc}
\usage{
deconvolute_bseqsc(
  bulk_gene_expression,
  signature = NULL,
  log = NULL,
  ...,
  verbose = FALSE
)
}
\arguments{
\item{bulk_gene_expression}{A matrix of bulk data. Rows are genes, columns are samples.
Row and column names need to be set.}

\item{signature}{The signature matrix. Rows are genes, columns are cell types.
If \code{NULL} then a built-in reference for pancreatic islet cell sub-population
is used (data PancreasIslet in bseqsc package).}

\item{log}{Logical that indicates if the data \code{x} is in in log-scale.
If \code{NULL}, then log scale is inferred by \link[xbioc:is_logscale]{xbioc::is_logscale}.}

\item{...}{Other arguments passed to \code{CIBERSORT}.}

\item{verbose}{Whether to produce an output on the console.}
}
\value{
A list including:
\item{coef}{The matrix of estimated proportions (cell type x samples).}
\item{stats}{The statistics computed by \code{CIBERSORT}.}
}
\description{
Deconvolution Analysis using BSEQ-sc
}
