% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CibersortX.R
\name{deconvolute_cibersort}
\alias{deconvolute_cibersort}
\title{Deconvolute with CibersortX}
\usage{
deconvolute_cibersort(
  bulk_gene_expression,
  signature,
  verbose = FALSE,
  input_dir = NULL,
  output_dir = NULL,
  display_extra_info = FALSE,
  label = "none",
  ...
)
}
\arguments{
\item{bulk_gene_expression}{A matrix or dataframe with the bulk data. Rows are genes, columns are samples.
Can also just be the filename of the bulk expression data in the correct format (generated by using this
method). The file needs to be in the folder "input_dir".}

\item{signature}{The signature matrix. Can also be the filename of the signature matrix in the correct
format (generated by using this method). The file needs to be in the folder "input_dir".}

\item{verbose}{Whether to output what CibersortX is doing.}

\item{input_dir}{The folder in which the input files can be found (or are created in). Default is a temporary directory.}

\item{output_dir}{The directory in which the output files are saved. Default is a temporary directory.}

\item{display_extra_info}{Whether to print the "P.value","Correlation" and "RMSE" calculated by CibersortX.}

\item{label}{The label which will be added to the file name. Default is "none", adding no label at all.}

\item{...}{Additional parameters supplied to the algorithm. Options are:
perm: <int> No. of permutations for p-value calculation (default: 0)
rmbatch_B_mode: <bool>  Run B-mode batch correction (default: FALSE)
rmbatch_S_mode: <bool>  Run S-mode batch correction (default: FALSE)
source_GEPs: <file_name>  Signature matrix GEPs for batch correction (default: sigmatrix)
qn: <bool>  Run quantile normalization (default: FALSE)
absolute: <bool>  Run absolute mode (default: FALSE)
abs_method: <char>  Pick absolute method ("sig.score" (default) or "no.sumto1")}
}
\value{
A matrix with the probabilities of each cell-type for each individual. Rows are individuals, columns are cell types.
}
\description{
Deconvolute with CibersortX
}
