% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{deconvolute_dwls}
\alias{deconvolute_dwls}
\title{Calculates the decomposition using the dwls algorithm}
\usage{
deconvolute_dwls(
  bulk_gene_expression,
  signature,
  dwls_submethod = c("DampenedWLS", "OLS", "SVR"),
  verbose = FALSE
)
}
\arguments{
\item{bulk_gene_expression}{An Expression Set containing bulk data.}

\item{signature}{The Signature matrix.}

\item{dwls_submethod}{Three alternative methods in DWLS: OLS, SVR, and DampenedWLS.}

\item{verbose}{Whether to produce an output on the console.}
}
\value{
A matrix of cell type proportion estimates with cell types as rows and individuals as
columns.
}
\description{
generates a reference profile based on single-cell data. Learns a transformation of bulk
expression based on observed single-cell proportions and performs NNLS regression on these
transformed values to estimate cell proportions.
}
