% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuSiC.R
\name{deconvolute_music}
\alias{deconvolute_music}
\title{MuSiC Deconvolution}
\usage{
deconvolute_music(
  bulk_gene_expression,
  signature_data,
  markers = NULL,
  clusters = "cellType",
  samples = "SubjectName",
  cell_size = NULL,
  ct_cov = FALSE,
  verbose = FALSE,
  iter.max = 1000,
  nu = 1e-04,
  eps = 0.01,
  centered = FALSE,
  normalize = FALSE
)
}
\arguments{
\item{bulk_gene_expression}{A matrix or dataframe with the bulk data. Rows
are genes, columns are samples.}

\item{signature_data}{Preprocessed single cell data from the build_model_music method}

\item{markers}{vector or list of gene names, default as NULL. If NULL, use all genes that
provided by both bulk and single cell dataset.}

\item{clusters}{character, the phenoData of single cell dataset used as clusters;}

\item{samples}{character,the phenoData of single cell dataset used as samples;}

\item{cell_size}{data.frame of cell sizes. 1st column contains the names of cell types, 2nd
column has the cell sizes per cell type. Default as NULL. If NULL, then estimate cell size from
data;}

\item{ct_cov}{logical. If TRUE, use the covariance across cell types;}

\item{verbose}{logical, default as TRUE.}

\item{iter.max}{numeric, maximum iteration number}

\item{nu}{regulation parameter, take care of weight when taking recipical}

\item{eps}{Thredshold of convergence}

\item{centered}{logic, substract avg of Y and D}

\item{normalize}{logic, divide Y and D by their standard deviation}
}
\value{
a list with elements:
\itemize{
\item Estimates of MuSiC
\item Estimates of NNLS
\item Weight of MuSiC
\item r.squared of MuSiC
\item Variance of MuSiC estimates
}
}
\description{
This function is to calculate the MuSiC deconvolution proportions
}
