% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{deconvolute_scaden}
\alias{deconvolute_scaden}
\title{Performs deconvolution with Scaden}
\usage{
deconvolute_scaden(
  signature,
  bulk_gene_expression,
  temp_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{signature}{Path to the model directory}

\item{bulk_gene_expression}{A matrix of bulk data. Rows are genes, columns are samples.
Row and column names need to be set.}

\item{temp_dir}{The temporary directory to use for the computations (optional)}

\item{verbose}{Whether to produce an output on the console (default: false).}
}
\value{
A matrix of cell type proportion estimates with cell types as rows and
individuals as columns.
}
\description{
Performs deconvolution with Scaden
}
