% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\docType{data}
\name{deconvolution_methods}
\alias{deconvolution_methods}
\title{List of supported deconvolution methods}
\format{
An object of class \code{character} of length 11.
}
\usage{
deconvolution_methods
}
\description{
The methods currently supported are
\code{AutoGeneS}, \code{Bisque}, \code{BSEQ-sc}, \code{CibersortX}, \code{CPM}, \code{DWLS}, \code{MOMF}, \code{MuSiC}, \code{Scaden},
\code{SCDC}
}
\details{
The object is a named vector. The names correspond to the display name of the method,
the values to the internal name.
}
\keyword{datasets}
