% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{deescape_blanks}
\alias{deescape_blanks}
\title{Removes the substitutions "$_$" and turns them back into blanks}
\usage{
deescape_blanks(string)
}
\arguments{
\item{string}{The string to be de-escaped}
}
\value{
The String with blanks
}
\description{
Removes the substitutions "$_$" and turns them back into blanks
}
