% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{deescape_special_chars}
\alias{deescape_special_chars}
\title{Removes the substitutions and turns them back into the special characters}
\usage{
deescape_special_chars(string)
}
\arguments{
\item{string}{The string to be de-escaped}
}
\value{
The String with special characters
}
\description{
Removes the substitutions and turns them back into the special characters
}
