% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper.R
\name{docker_available}
\alias{docker_available}
\title{Docker availability check}
\usage{
docker_available()
}
\value{
A boolean value whether docker is available on the system
}
\description{
Docker availability check
}
