% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{docker_connectable}
\alias{docker_connectable}
\title{Docker connectability check}
\usage{
docker_connectable()
}
\value{
A boolean value whether it is possible to connect to docker
}
\description{
Docker connectability check
}
