% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{escape_blanks}
\alias{escape_blanks}
\title{Removes blanks by substituting them with "$_$" which should not be used naturally}
\usage{
escape_blanks(string)
}
\arguments{
\item{string}{The string to be escaped}
}
\value{
The String without blanks
}
\description{
Removes blanks by substituting them with "$_$" which should not be used naturally
}
