% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{escape_special_chars}
\alias{escape_special_chars}
\title{Removes special characters by substituting them with unique string which should not be used naturally}
\usage{
escape_special_chars(string)
}
\arguments{
\item{string}{The string to be escaped}
}
\value{
The String without special characters
}
\description{
Removes special characters by substituting them with unique string which should not be used naturally
}
