% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneS.R
\name{extract_signature_autogenes}
\alias{extract_signature_autogenes}
\title{Extraction of the signature matrix from autogenes .pickle files}
\usage{
extract_signature_autogenes(
  autogenes_pickle_path,
  single_cell_object,
  cell_type_annotations
)
}
\arguments{
\item{autogenes_pickle_path}{The path of the .pickle file generated
by autogenes, exactly as returned by the \code{build_model_autogenes} function}

\item{single_cell_object}{The matrix with the single-cell data used to build the signature.
Rows are genes, columns are samples. Row and column names need to be set.}

\item{cell_type_annotations}{The vector of the cell type annotations used to
build the signature. Has to be in the same order as the samples in single_cell_object.}
}
\value{
The signature matrix.
}
\description{
Extraction of the signature matrix from autogenes .pickle files
}
