% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{find_dampening_constant}
\alias{find_dampening_constant}
\title{Finding a dampening constant for the weights using cross-validation}
\usage{
find_dampening_constant(S, B, gold_standard)
}
\arguments{
\item{S}{Signature matrix}

\item{B}{One column of the bulk data matrix, so one sample}

\item{gold_standard}{The solution found with OLS}
}
\value{
The dampening constant (integer)
}
\description{
Finding a dampening constant for the weights using cross-validation
}
