% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{findDampeningConstant}
\alias{findDampeningConstant}
\title{Finding a dampening constant for the weights using cross-validation}
\usage{
findDampeningConstant(S, B, goldStandard, verbose = FALSE)
}
\arguments{
\item{S}{Signature matrix}

\item{B}{One column of the bulk data matrix, so one sample}

\item{goldStandard}{The solution found with OLS}

\item{verbose}{Whether to produce an output on the console}
}
\value{
The dampening constant (integer)
}
\description{
Finding a dampening constant for the weights using cross-validation
}
