% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{fix_dependencies}
\alias{fix_dependencies}
\title{Manage python dependencies
according to: https://rstudio.github.io/reticulate/articles/python_dependencies.html#manual-configuration}
\description{
Manage python dependencies
according to: https://rstudio.github.io/reticulate/articles/python_dependencies.html#manual-configuration
}
