% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CibersortX.R
\name{get_cell_fractions_options}
\alias{get_cell_fractions_options}
\title{Creation of the options of the "get cell fractions" docker command}
\usage{
get_cell_fractions_options(
  sigmatrix,
  mixture,
  perm = 0,
  label = "none",
  rmbatch_B_mode = FALSE,
  rmbatch_S_mode = FALSE,
  source_GEPs = sigmatrix,
  qn = FALSE,
  absolute = FALSE,
  abs_method = "sig.score"
)
}
\arguments{
\item{sigmatrix}{The filename of the signature matrix.}

\item{mixture}{The filename of the bulk expression data.}

\item{perm}{<int> No. of permutations for p-value calculation (default: 0).}

\item{label}{The label which will be added to the file name. Default is "none", adding no label
at all.}

\item{rmbatch_B_mode}{<bool>  Run B-mode batch correction (default: FALSE).}

\item{rmbatch_S_mode}{<bool>  Run S-mode batch correction (default: FALSE).}

\item{source_GEPs}{<file_name>  Signature matrix GEPs for batch correction (default: sigmatrix).}

\item{qn}{<bool>  Run quantile normalization (default: FALSE).}

\item{absolute}{<bool>  Run absolute mode (default: FALSE).}

\item{abs_method}{<char>  Pick absolute method ("sig.score" (default) or "no.sumto1").}
}
\value{
A string in the correct format for the docker command, containing all parameters.
}
\description{
Creation of the options of the "get cell fractions" docker command
}
