% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{getSingleCellExpressionSet}
\alias{getSingleCellExpressionSet}
\title{Generating a single cell expression set}
\usage{
getSingleCellExpressionSet(single_cell_matrix, sampleNames, genes, cellTypes)
}
\arguments{
\item{single_cell_matrix}{The single-cell matrix. Rows are genes, columns are samples.}

\item{sampleNames}{A vector of the names of the samples, basically colnames(single_cell_matrix).}

\item{genes}{A vector of the names of the genes, basically rownames(single_cell_matrix).}

\item{cellTypes}{A Vector of the cell type annotations. Has to be in the same order as the samples in single_cell_object.}
}
\value{
A Biobase::ExpressionSet of the input data.
}
\description{
Generating a single cell expression set
}
