% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_config.R
\name{init_python}
\alias{init_python}
\title{Initiates python environment}
\usage{
init_python(python = NULL)
}
\arguments{
\item{python}{(optional) If own python should be used please indicate it's binaries}
}
\description{
Initiates python environment
}
