% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneS.R
\name{install_autogenes}
\alias{install_autogenes}
\title{Install AutoGeneS}
\usage{
install_autogenes()
}
\description{
Install AutoGeneS into a previously defined python environment.
A python environment can be defined by set_virtualenv() or set_python() command.
Alternatively a new environment can be created via create_virtualenv() method.
}
