% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectanglepy.R
\name{install_rectangle_python}
\alias{install_rectangle_python}
\title{Install Python dependencies for Rectangle}
\usage{
install_rectangle_python()
}
\value{
Invisibly NULL. Called for its side effect of setting up the conda environment.
}
\description{
Creates the r-omnideconv-rectangle conda environment with Python 3.10 and
installs rectanglepy. This is called automatically on first use but can be
run in advance to avoid the setup delay.
}
