% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{install_scaden}
\alias{install_scaden}
\title{Install Scaden}
\usage{
install_scaden(method = "auto", conda = "auto")
}
\arguments{
\item{conda}{}
}
\value{

}
\description{
Install Scaden into a previously defined python environment.
}
