% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpy.R
\name{install_scanpy}
\alias{install_scanpy}
\title{Install scanpy into python environment.}
\usage{
install_scanpy(method = "auto", conda = "auto")
}
\arguments{
\item{conda}{}
}
\value{

}
\description{
Install scanpy into python environment.
}
