% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{matrix_to_singlecellexperiment}
\alias{matrix_to_singlecellexperiment}
\title{Convert count matrix to SingleCellExperiment}
\usage{
matrix_to_singlecellexperiment(
  matrix,
  cell_labels,
  named_metadata_list = list()
)
}
\arguments{
\item{matrix}{Matrix of single cell data (rows: genes, columns: cells), the row names of the matrix should be labeled with gene labels}

\item{cell_labels}{Annotation vector for celltypes (e.g columns of matrix)}

\item{named_metadata_list}{A named list containing unstructured metadata information (optional)}
}
\value{
SingleCellExperiment
}
\description{
Convert count matrix to SingleCellExperiment
}
