% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{normalize_deconv_results}
\alias{normalize_deconv_results}
\title{Normalize deconvolution result}
\usage{
normalize_deconv_results(deconv_result)
}
\arguments{
\item{deconv_result}{The original deconvolution result}
}
\value{
A matrix with the rowsums of one and no negative values
}
\description{
Normalize deconvolution result
}
