% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneS.R
\name{null_transformer}
\alias{null_transformer}
\title{Null transformer for glue: Replaces NULL values with something else}
\usage{
null_transformer(str = "NULL")
}
\arguments{
\item{str}{The replacement string}
}
\value{
A function used for the transformation
}
\description{
Null transformer for glue: Replaces NULL values with something else
}
