% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{omnideconvstartup}
\alias{omnideconvstartup}
\title{Manage python dependencies
according to: https://rstudio.github.io/reticulate/articles/python_dependencies.html#manual-configuration}
\description{
Manage python dependencies
according to: https://rstudio.github.io/reticulate/articles/python_dependencies.html#manual-configuration
}
