% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_deconvolution}
\alias{plot_deconvolution}
\title{Plot Deconvolution results}
\usage{
plot_deconvolution(deconvolutions, plotMethod, facets, palette = "Set1")
}
\arguments{
\item{deconvolutions}{A named list of deconvolution results}

\item{plotMethod}{Type of plot to be rendered  ("bar", "jitter", "scatter", "box", "heatmap")}

\item{facets}{Variable for grouping the plots ("method", "cell_type", "sample")}

\item{palette}{RColorBrewer palette name (optional), standard = "Set1"}
}
\value{
ggplot rendered by plotly for interactivity
}
\description{
Plot Deconvolution results
}
\examples{
data("single_cell_data_1")
data("cell_type_annotations_1")
data("batch_ids_1")
data("bulk")

deconvolution <- omnideconv::deconvolute(bulk, NULL, "bisque",
                                         single_cell_data_1,
                                         cell_type_annotations_1,
                                         batch_ids_1)
deconvolution <- list(deconvolution)
names(deconvolution) <- "bisque"
omnideconv::plot_deconvolution(deconvolution, "bar", "method", "Spectral")

}
