% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotDeconvResult}
\alias{plotDeconvResult}
\title{Plot deconvolution results as a barplot}
\usage{
plotDeconvResult(deconv_result, method_name = "", file_name = NULL)
}
\arguments{
\item{deconv_result}{result from deconvolution()}

\item{method_name}{(optional) title of plot is set to the name of the method used for deconvolution}

\item{file_name}{(optional) plot is saved in this file}
}
\value{
the ggplot object
}
\description{
Plot deconvolution results as a barplot
}
\examples{
plotDeconvResult(immunedeconv2::deconvolute(bulk, immunedeconv2::build_model(single_cell_data, cell_type_annotations, "bisque"), "bisque", single_cell_data, cell_type_annotations), "Bisque")
}
