% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_config.R
\name{python_available}
\alias{python_available}
\title{Checks if python is available in environment}
\usage{
python_available()
}
\value{
boolean
}
\description{
Checks if python is available in environment
}
