% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneS.R
\name{quote_string}
\alias{quote_string}
\title{Surrounds a string with quotes}
\usage{
quote_string(string)
}
\arguments{
\item{string}{The string to be quoted}
}
\value{
The quoted string
}
\description{
Surrounds a string with quotes
}
