% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RefData}
\alias{RefData}
\title{Gold standard measurements with FACS from Hoek et al.}
\format{
matrix with samples x cell types
}
\source{
Hoek et al. (2015), PLOS ONE, https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0118528
}
\usage{
RefData
}
\description{
A dataset of cell-type fractions of eight RNA-seq samples measured with FACS
}
\keyword{datasets}
