% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\docType{data}
\name{required_packages}
\alias{required_packages}
\title{The dependencies for each method}
\format{
An object of class \code{list} of length 12.
}
\usage{
required_packages
}
\description{
The dependencies for each method
}
\keyword{datasets}
