% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{save_as_h5ad}
\alias{save_as_h5ad}
\title{Save as .h5ad}
\usage{
save_as_h5ad(single_cell_object, cell_type_annotations)
}
\arguments{
\item{single_cell_object}{A matrix or dataframe with the single-cell data. Rows are genes,
columns are samples. Row and column names need to be set. Alternatively a SingleCellExperiment
or an AnnData object can be provided. In that case, note that cell-type labels need to be
indicated either directly providing a vector (cell_type_annotations) or by indicating the
column name that indicates the cell-type labels (cell_type_column_name). (Anndata: obs object,
SingleCellExperiment: colData object)}

\item{cell_type_annotations}{A Vector of the cell type annotations. Has to be in the same order
as the samples in single_cell_object}
}
\value{
The path to the saved .h5ad file
}
\description{
Transform a single_cell_matrix into an anndata object and saves it
}
