% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\name{scaden}
\alias{scaden}
\title{Scaden model creation and deconvolution}
\usage{
scaden(
  single_cell_object,
  cell_type_annotations,
  gene_labels,
  bulk_data,
  anndata_object = NULL,
  verbose = F,
  ...
)
}
\arguments{
\item{single_cell_object}{A matrix or dataframe with the single-cell expression data. Rows are cells, columns are genes.}

\item{cell_type_annotations}{Vector of celltype labels. Same order as rows in single_cell_object.}

\item{gene_labels}{Vector of gene symbols. Same order as columns in single_cell_object.}

\item{bulk_data}{Matrix or dataframe of bulk RNA expression. Rows are genes, columns are samples.}

\item{anndata_object}{Single cell expression in AnnData format. If provided, single_cell_object, cell_type_annotations and gene_labels are ignored.}

\item{...}{Parameters that can be forwarded to Scaden}
}
\value{

}
\description{
Scaden model creation and deconvolution
}
\examples{
# Calling scaden with single cell expression matrix + annotations:
# cell_type_proportions <- scaden(single_cell_matrix ,cell_type_annotation ,gene_labels ,bulk_data)

# Calling scaden with .h5ad file:
# cell_type_proportions <- scaden(anndata_object = h5ad_file, bulk_data = bulk_data)

# Caution: The process step of scaden fails if inputting a single_cell experiment that is log-transformed.

}
