% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{scaden_build_model_from_h5ad}
\alias{scaden_build_model_from_h5ad}
\title{Builds Scaden model from training data in .h5ad format.}
\usage{
scaden_build_model_from_h5ad(
  training_data_h5ad,
  bulk_data,
  model_path = NULL,
  verbose = F,
  ...
)
}
\arguments{
\item{training_data_h5ad}{Training data in .h5ad format.}

\item{bulk_data}{Matrix of bulk RNA expression. (genes x samples)}

\item{model_path}{Path where model directory should be created.}
}
\value{

}
\description{
Builds Scaden model from training data in .h5ad format.
}
