% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{scaden_checkload}
\alias{scaden_checkload}
\title{Checks if scaden is installed.}
\usage{
scaden_checkload(python = NULL)
}
\arguments{
\item{python}{The python env}
}
\description{
If it is available, the python module is imported.
}
