% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{scaden_simulate}
\alias{scaden_simulate}
\title{Simulates training data from scRNA data}
\usage{
scaden_simulate(
  celltype_labels,
  gene_labels,
  count_data,
  cells = 100,
  samples = 1000,
  verbose = F
)
}
\arguments{
\item{celltype_labels}{Vector of celltype labels. Order corresponds to rows in count_data matrix.}

\item{gene_labels}{Vector of gene labels. Order corresponds to columns in count_data matrix.}

\item{count_data}{Matrix or dataframe of scRNA data. Rows=cells and columns=genes}
}
\value{

}
\description{
Simulates training data from scRNA data
}
