% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{scaden_simulate_example}
\alias{scaden_simulate_example}
\title{Simulates example Data provided by Scaden}
\usage{
scaden_simulate_example(example_data_path = NULL, verbose = F)
}
\arguments{
\item{example_data_path}{Path to where example data should be saved. (directory)}

\item{verbose}{Defines verbosity of function call (default: false)}
}
\value{
List with list$simulated_h5ad =  example training data
and list$bulk = example bulk data.
}
\description{
Used for testing.
}
