% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{SCEs_are_identical}
\alias{SCEs_are_identical}
\title{Check if two SingleCellExperiments are identical}
\usage{
SCEs_are_identical(a, b)
}
\arguments{
\item{a}{SingleCellExperiments}

\item{b}{SingleCellExperiments}
}
\value{
boolean
}
\description{
Check if two SingleCellExperiments are identical
}
