% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORTx.R
\name{set_cibersortx_credentials}
\alias{set_cibersortx_credentials}
\title{Setting the CIBERSORTx Credentials}
\usage{
set_cibersortx_credentials(email, token)
}
\arguments{
\item{email}{The email which was used to register on the website.}

\item{token}{The token provided by the website.}
}
\description{
Setting the CIBERSORTx Credentials
}
\examples{
set_cibersortx_credentials("max.mustermann@online.de", "7250fe7ajc322eeta192d163d62e6399")
}
