% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_config.R
\name{set_python}
\alias{set_python}
\title{Defines python path.}
\usage{
set_python(path)
}
\arguments{
\item{path}{Path to the Python binaries.}
}
\value{

}
\description{
Python>=3.7 is needed for scanpy and scaden to work.
}
