% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_config.R
\name{set_virtualenv}
\alias{set_virtualenv}
\title{Defines a virtual environment that should be used.}
\usage{
set_virtualenv(env_path)
}
\arguments{
\item{env_path}{Path to virtual environment.}
}
\value{

}
\description{
This environment needs to contain python3 for scanpy and scaden to work.
}
