% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setup_apptainer_container}
\alias{setup_apptainer_container}
\title{Setup of the apptainer container}
\usage{
setup_apptainer_container(container_path = NULL)
}
\arguments{
\item{container_path}{the path where the apptainer .sif file should be stored (optional)
If the file 'fractions_latest.sif' is already present, it will be used}
}
\value{
the path to the apptainer container
}
\description{
Setup of the apptainer container
}
