% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setup_singularity_container}
\alias{setup_singularity_container}
\title{Setup of the singularity container}
\usage{
setup_singularity_container(container_path = NULL)
}
\arguments{
\item{container_path}{the path where the singularity .sif file should be stored (optional)
If the file 'fractions_latest.sif' is already present, it will be used}
}
\value{
the path to the singularity container
}
\description{
Setup of the singularity container
}
