% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_cell_data_3}
\alias{single_cell_data_3}
\title{Example single-cell RNA-seq dataset Number 3 from Hao et al.}
\format{
A matrix with 19131 genes as rows and 2400 cells as columns
}
\source{
Hao et al. (2021): Integrated analysis of multimodal single-cell data.
In Cell 184 (13), 3573-3587.e29. DOI: 10.1016/j.cell.2021.04.048.
}
\usage{
single_cell_data_3
}
\description{
A subset of raw counts of single-cell RNA-seq gene expression, containing 2400 cells.
his contains no cells which are in the other two subsets and all genes with only 0's in
all cells are removed
}
\keyword{datasets}
