% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_cell_data_small}
\alias{single_cell_data_small}
\title{Small example sigle-cell RNA-seq dataset from Maynard et al.}
\format{
matrix with genes x cells
}
\source{
Maynard et al. (2020), Cell,
https://www.sciencedirect.com/science/article/pii/S0092867420308825
}
\usage{
single_cell_data_small
}
\description{
A small subset of TPM-normalized single-cell RNA-seq gene expression, containing 50 cells with
types "T cell CD4", "T cell CD8", "B cell", "Monocyte conventional", "NK cell" and only 100 genes
}
\keyword{datasets}
