% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{SingleCellExperiment_to_Anndata}
\alias{SingleCellExperiment_to_Anndata}
\title{Convert SingleCellExperiment to AnnData object}
\usage{
SingleCellExperiment_to_Anndata(sce, X_name = NULL)
}
\arguments{
\item{sce}{object of type SingleCellExperiment}

\item{X_name}{Name of the assay in the SingleCellExperiment to interpret as the X matrix of the AnnData object (default: first in assay list)}
}
\value{
AnnData object
}
\description{
Convert SingleCellExperiment to AnnData object
}
