% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{solve_dampened_wls}
\alias{solve_dampened_wls}
\title{Solver using WLS with weights dampened by a certain dampening constant}
\usage{
solve_dampened_wls(S, B, verbose = FALSE)
}
\arguments{
\item{S}{Signature matrix}

\item{B}{One column of the bulk data matrix, so one sample}

\item{verbose}{Whether to produce an output on the console.}
}
\value{
A vector with the cell type numbers for the sample
}
\description{
Solver using WLS with weights dampened by a certain dampening constant
}
