% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{solve_dampened_wlsj}
\alias{solve_dampened_wlsj}
\title{Solve WLS given a dampening constant}
\usage{
solve_dampened_wlsj(S, B, gold_standard, j, verbose = FALSE)
}
\arguments{
\item{S}{Signature matrix}

\item{B}{One column of the bulk data matrix, so one sample}

\item{gold_standard}{The average of all the solutions so far}

\item{j}{The dampening constant}

\item{verbose}{Whether to produce an output on the console.}
}
\value{
A vector with the cell type numbers for the sample
}
\description{
Solve WLS given a dampening constant
}
