% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{solveDampenedWLSj}
\alias{solveDampenedWLSj}
\title{Solve WLS given a dampening constant}
\usage{
solveDampenedWLSj(S, B, goldStandard, j, verbose = FALSE)
}
\arguments{
\item{S}{Signature matrix}

\item{B}{One column of the bulk data matrix, so one sample}

\item{goldStandard}{The average of all the solutions so far}

\item{j}{The dampening constant}

\item{verbose}{Whether to print output}
}
\value{
A vector with the cell type numbers for the sample
}
\description{
Solve WLS given a dampening constant
}
