% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{solveOLS}
\alias{solveOLS}
\title{Solver using OLS, constrained such that cell type numbers>0}
\usage{
solveOLS(S, B, verbose = FALSE)
}
\arguments{
\item{S}{Signature matrix}

\item{B}{One column of the bulk data matrix, so one sample}

\item{verbose}{Whether to produce an output on the console.}
}
\value{
A vector with the cell type proportions for the sample
}
\description{
Solver using OLS, constrained such that cell type numbers>0
}
