% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{solveOLSInternal}
\alias{solveOLSInternal}
\title{Solver using OLS, constrained such that cell type numbers>0}
\usage{
solveOLSInternal(S, B, verbose = FALSE)
}
\arguments{
\item{S}{Signature matrix}

\item{B}{One column of the bulk data matrix, so one sample}

\item{verbose}{Whether to print output}
}
\value{
A vector with the cell type numbers for the sample
}
\description{
Returns absolute numbers, not proportions
}
