% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{solveSVR}
\alias{solveSVR}
\title{Solver using SVR}
\usage{
solveSVR(S, B)
}
\arguments{
\item{S}{Signature matrix}

\item{B}{One column of the bulk data matrix, so one sample}
}
\value{
A vector with the cell type proportions for the sample
}
\description{
Solver using SVR
}
