% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CibersortX.R
\name{transform_and_save_bulk}
\alias{transform_and_save_bulk}
\title{Creation of the bulk data file in the CibersortX required format}
\usage{
transform_and_save_bulk(bulk, path, verbose = FALSE)
}
\arguments{
\item{bulk}{The bulk data as a matrix.}

\item{path}{The folder in which the file should be saved.}

\item{verbose}{Whether to create any output.}
}
\value{
The path to the bulk data file
}
\description{
Creation of the bulk data file in the CibersortX required format
}
