% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CibersortX.R
\name{transform_and_save_single_cell}
\alias{transform_and_save_single_cell}
\title{Creation of the single cell data file in the CIBERSORTx required format}
\usage{
transform_and_save_single_cell(sc_matrix, cell_types, path, verbose = FALSE)
}
\arguments{
\item{sc_matrix}{The single cell data as a matrix.}

\item{cell_types}{A Vector of the cell type annotations. Has to be in the same order as the
samples in single_cell_object.}

\item{path}{The folder in which the file should be saved.}

\item{verbose}{Whether to produce an output on the console.}
}
\value{
The path to the single cell data file
}
\description{
Creation of the single cell data file in the CIBERSORTx required format
}
