% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneS.R
\name{transform_boolean}
\alias{transform_boolean}
\title{Transforms the R TRUE and FALSE into Pythons True and False}
\usage{
transform_boolean(bool)
}
\arguments{
\item{bool}{The original boolean}
}
\value{
The python readable boolean
}
\description{
Transforms the R TRUE and FALSE into Pythons True and False
}
