% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{verbose_wrapper}
\alias{verbose_wrapper}
\title{A wrapper function whether to suppress messages}
\usage{
verbose_wrapper(verbose)
}
\arguments{
\item{verbose}{Whether to produce an output on the console.}
}
\value{
A function which will suppress messages or not, depending on the verbose parameter
}
\description{
A wrapper function whether to suppress messages
}
