% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnData.R
\name{write_anndata}
\alias{write_anndata}
\title{Write anndata object (.h5ad format)}
\usage{
write_anndata(data, path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Write anndata object (.h5ad format)
}
