% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{anndata_is_identical}
\alias{anndata_is_identical}
\title{Checks if anndata package is loaded}
\usage{
anndata_is_identical(a, b)
}
\description{
If called and python environment is not set up, this is realized. Else, it checks if the anndata
package is loaded, and if not, it does this.
}
