% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{anndata_to_singlecellexperiment}
\alias{anndata_to_singlecellexperiment}
\title{Convert AnnData to SingleCellExperiment}
\usage{
anndata_to_singlecellexperiment(ad)
}
\arguments{
\item{ad}{AnnData object}
}
\value{
SingleCellObject
}
\description{
Convert AnnData to SingleCellExperiment
}
