% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\name{build_model}
\alias{build_model}
\title{Building the signature matrix}
\usage{
build_model(
  single_cell_object,
  cell_type_annotations,
  method = deconvolution_methods,
  bulk_gene_expression = NULL,
  ...
)
}
\arguments{
\item{single_cell_object}{A matrix or dataframe with the single-cell data. Rows are genes, columns are samples. Row and column names need to be set.}

\item{cell_type_annotations}{A Vector of the cell type annotations. Has to be in the same order as the samples in single_cell_object}

\item{method}{A string specifying the method.
Supported methods are \"bisque\", \"momf\", \"dwls\", \"...\"}

\item{bulk_gene_expression}{A matrix of bulk data. Rows are genes, columns are samples. Necessary for MOMF, defaults to NULL.Row and column names need to be set.}

\item{...}{Additional parameters, passed to the algorithm used.}
}
\value{
The signature matrix. Rows are genes, columns are cell types.
}
\description{
The single_cell_object is expected to have rownames() and colnames()
}
