% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\name{build_model}
\alias{build_model}
\title{Building the signature matrix}
\usage{
build_model(
  single_cell_object,
  cell_type_annotations,
  method = deconvolution_methods,
  ...
)
}
\arguments{
\item{single_cell_object}{A matrix or dataframe with the single-cell data. Rows are genes, columns are samples.}

\item{cell_type_annotations}{A Vector of the cell type annotations. Has to be in the same order as the samples in single_cell_object}

\item{method}{A string specifying the method.
Supported methods are \"bisque\", \"momf\", \"...\"}

\item{...}{Additional parameters, passed to the algorithm used.}
}
\value{
The signature matrix. Rows are genes, columns are cell types.
}
\description{
Building the signature matrix
}
