% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CibersortX.R
\name{build_model_cibersortx}
\alias{build_model_cibersortx}
\title{Signature matrix creation with CibersortX}
\usage{
build_model_cibersortx(
  single_cell_object,
  cell_type_annotations,
  verbose = FALSE,
  input_dir = NULL,
  output_dir = NULL,
  display_heatmap = FALSE,
  k_max = 999,
  ...
)
}
\arguments{
\item{single_cell_object}{A matrix with the single-cell data. Rows are genes, columns are
samples. Row and column names need to be set.}

\item{cell_type_annotations}{A vector of the cell type annotations. Has to be in the same order
as the samples in single_cell_object.}

\item{verbose}{Whether to produce an output on the console.}

\item{input_dir}{The directory in which the input files can be found (or are created in).
Default is a temporary directory.}

\item{output_dir}{The directory in which the output files are saved. Default is a temporary
directory.}

\item{display_heatmap}{Whether to display the generated heatmap.}

\item{k_max}{Maximum condition number (default: 999). Will be added to the file name.}

\item{...}{Additional parameters supplied to the algorithm. Options are:
\itemize{
\item G_min <int> Minimum number of genes per cell type in sig. matrix (default: 300).
\item G_max <int> Maximum number of genes per cell type in sig. matrix (default: 500).
\item q_value <int> Q-value threshold for differential expression (default: 0.01).
\item filter <bool> Remove non-hematopoietic genes (default: FALSE).
\item remake <bool> Remake signature gene matrix (default: False).
\item replicates <int> Number of replicates to use for building scRNAseq reference file
(default: 5).
\item sampling <float> Fraction of available single cell GEPs selected using random sampling
(default: 0.5).
\item fraction <float> Fraction of cells of same identity showing evidence of expression
(default: 0.75).
}}
}
\value{
The signature matrix. Rows are genes, columns are cell types.
}
\description{
Signature matrix creation with CibersortX
}
