% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWLS.R
\name{build_model_dwls}
\alias{build_model_dwls}
\title{Signature matrix creation with DWLS using genes identified by a differential analysis}
\usage{
build_model_dwls(
  single_cell_object,
  cell_type_annotations,
  dwls_method = c("mast", "seurat"),
  path = NULL,
  verbose = FALSE,
  diff_cutoff = 0.5,
  pval_cutoff = 0.01
)
}
\arguments{
\item{single_cell_object}{A matrix with the single-cell data. Rows are genes, columns are
samples. Row and column names need to be set.}

\item{cell_type_annotations}{A vector of the cell type annotations. Has to be in the same order
as the samples in single_cell_object.}

\item{dwls_method}{The method used to create the signature matrix. Options are "mast" and "seurat"}

\item{path}{The path where the generated files will be saved. If path=NULL, the generated files
will be discarded.}

\item{verbose}{Whether to produce an output on the console.}

\item{diff_cutoff}{Cutoff to determine the FC-limit. How low can the lowest fold change be to
still be considered differentially expressed?}

\item{pval_cutoff}{Cutoff to determine the pVal-limit. How high can the highest p-Value be to
still be considered statistically significant?}
}
\value{
The signature matrix. Rows are genes, columns are cell types.
}
\description{
Signature matrix creation with DWLS using genes identified by a differential analysis
}
