% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCDC.R
\name{build_model_scdc}
\alias{build_model_scdc}
\title{No model is build as SCDC does both steps in one.}
\usage{
build_model_scdc()
}
\value{
NULL.
}
\description{
Please use the deconvolute method with your single cell and bulk rna seq data to use SCDC.
}
