% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\name{deconvolute}
\alias{deconvolute}
\title{Deconvolution}
\usage{
deconvolute(
  bulk_gene_expression,
  signature,
  method = deconvolution_methods,
  single_cell_object = NULL,
  cell_type_annotations = NULL,
  verbose = T,
  ...
)
}
\arguments{
\item{bulk_gene_expression}{A matrix or dataframe with the bulk data. Rows are genes, columns are samples.}

\item{signature}{The signature matrix.}

\item{method}{A string specifying the method.
Supported methods are \"bisque\", \"momf\", \"dwls\", \"...\"}

\item{single_cell_object}{Needed for deconvolution with MOMF and Bisque. Defaults to NULL.}

\item{cell_type_annotations}{Needed for deconvolution with Bisque. Defaults to NULL.}

\item{verbose}{Whether the algorithms should print out what they are doing.}

\item{...}{Additional parameters, passed to the algorithm used.}
}
\value{
A matrix with the probabilities of each cell-type for each individual. Rows are individuals, columns are cell types.
}
\description{
Deconvolution
}
