% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneS.R
\name{deconvolute_autogenes}
\alias{deconvolute_autogenes}
\title{Deconvolution Analysis using AutoGeneS}
\usage{
deconvolute_autogenes(
  bulk_gene_expression,
  signature,
  single_cell_object,
  verbose = FALSE,
  method = "KL",
  ...
)
}
\arguments{
\item{bulk_gene_expression}{Dataframe or matrix of bulk RNA-seq data (genes x individuals)}

\item{signature}{Signature Matrix (genes x individuals from scRNA-seq)}

\item{single_cell_object}{scRNA-seq Object (genes x cells)}

\item{verbose}{Whether the algorithm should print out what it is doing.}

\item{method}{Determines which divergence to use. Options: Kullback-Leibler "KL", Itakura-Saito "IS". Defaults to "KL"}

\item{...}{additional parameters}
}
\value{
cell proportion matrix
}
\description{
Deconvolution Analysis using AutoGeneS
}
