% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneS.R
\name{deconvolute_autogenes}
\alias{deconvolute_autogenes}
\title{Deconvolution Analysis using AutoGeneS}
\usage{
deconvolute_autogenes(
  bulk_gene_expression,
  signature,
  model = c("nusvr", "nnls", "linear"),
  nu = 0.5,
  C = 0.5,
  normalize_results = TRUE,
  kernel = "linear",
  degree = 3,
  gamma = "scale",
  coef0 = 0,
  shrinking = TRUE,
  tol = 0.001,
  cache_size = 200,
  max_iter = -1,
  weights = NULL,
  index = NULL,
  close_to = NULL,
  verbose = FALSE
)
}
\arguments{
\item{bulk_gene_expression}{A matrix of bulk data. Rows are genes, columns are samples.
Row and column names need to be set.}

\item{signature}{Path to a .pickle file, created with the build_model method of AutoGeneS.}

\item{model}{Regression model. Available options: NuSVR ("nusvr"), non-negative least
squares("nnls") and linear model ("linear").}

\item{nu}{Nu parameter for NuSVR.}

\item{C}{C parameter for NuSVR.}

\item{normalize_results}{wether to normalize results according to the regression model used.
Default is TRUE}

\item{kernel}{Kernel parameter for NuSVR.}

\item{degree}{Degree parameter for NuSVR.}

\item{gamma}{Gamma parameter for NuSVR.}

\item{coef0}{Coef0 parameter for NuSVR.}

\item{shrinking}{Shrinking parameter for NuSVR.}

\item{tol}{Tol parameter for NuSVR.}

\item{cache_size}{Cache_size parameter for NuSVR.}

\item{max_iter}{Max_iter parameter for NuSVR.}

\item{weights}{Select Solution: Weights with which to weight the objective values. For example,
(-1,2) will minimize the first objective and maximize the the second (with more weight).}

\item{index}{Select Solution: If one int is passed, return pareto[index] If two ints are passed,
the first is an objective (0 for the first). The second is the nth element if the solutions
have been sorted by the objective in ascending order. For example, (0,1) will return the
solution that has the second-lowest value in the first objective. (1,-1) will return the
solution with the highest value in the second objective.}

\item{close_to}{Select Solution: Select the solution whose objective value is close to a certain
value. Assumes (objective,value). For example, (0,100) will select the solution whose value
for the first objective is closest to 100.}

\item{verbose}{Whether to produce an output on the console.}
}
\value{
A list with two elements: 'proportions' is the matrix of cell proportions and
'genes_used' is a vector containing the names of the genes used for the deconvolution, what
is called "solution" by AutoGeneS.
}
\description{
Deconvolution Analysis using AutoGeneS
}
