% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSEQ_sc.R
\name{deconvolute_bseqsc}
\alias{deconvolute_bseqsc}
\title{Deconvolution Analysis using BSEQ-sc}
\usage{
deconvolute_bseqsc(
  bulk_gene_expression,
  signature = NULL,
  log = NULL,
  ...,
  verbose = FALSE
)
}
\arguments{
\item{bulk_gene_expression}{Dataframe or matrix of bulk RNA-seq data (genes x individuals)}

\item{signature}{Signature Matrix (genes x individuals from scRNA-seq)
If \code{NULL} then a built-in reference for pancreatic islet cell sub-population
is used (data PancreasIslet in bseqsc package).}

\item{log}{logical that indicates if the data \code{x} is in in log-scale.
If \code{NULL}, then log scale is inferred by \link[xbioc:is_logscale]{xbioc::is_logscale}.}

\item{...}{other arguments passed to \code{CIBERSORT}.}

\item{verbose}{logical that toggles log messages.}
}
\value{
a list with elements:
\itemize{
\item coef: matrix of estimated proportions (cell type x samples)
\item stats: statistics computed by \code{CIBERSORT}
}

cell proportion matrix
}
\description{
Deconvolution Analysis using BSEQ-sc
}
