% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOMF.R
\name{deconvolute_momf}
\alias{deconvolute_momf}
\title{Deconvolution Analysis using MOMF (via Nonnegative Factorization)}
\usage{
deconvolute_momf(
  bulk_gene_expression,
  signature,
  single_cell_object,
  verbose = FALSE,
  method = "KL",
  ...
)
}
\arguments{
\item{bulk_gene_expression}{A matrix of bulk data. Rows are genes, columns are samples.
Row and column names need to be set.}

\item{signature}{The signature matrix. Rows are genes, columns are cell types.}

\item{single_cell_object}{A matrix with the single-cell data. Rows are genes, columns are
samples. Row and column names need to be set.}

\item{verbose}{Whether to produce an output on the console.}

\item{method}{Determines which divergence to use. Options: Kullback-Leibler "KL",
Itakura-Saito "IS". Defaults to "KL"}

\item{...}{additional parameters}
}
\value{
cell proportion matrix
}
\description{
Deconvolution Analysis using MOMF (via Nonnegative Factorization)
}
