% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{deconvolute_scaden}
\alias{deconvolute_scaden}
\title{Performs deconvolution}
\usage{
deconvolute_scaden(model, bulk_data, verbose = FALSE)
}
\arguments{
\item{model}{Path to the model directory}

\item{bulk_data}{Matrix of bulk RNA expression. (genes x samples)}

\item{verbose}{Defines verbosity of function call (default: false)}
}
\value{
Cell type fractions per sample.
}
\description{
Performs deconvolution
}
