% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCDC.R
\name{deconvolute_scdc}
\alias{deconvolute_scdc}
\title{SCDC Deconvolution}
\usage{
deconvolute_scdc(
  bulk_gene_expression,
  single_cell_object,
  cell_type_annotations,
  batch_ids,
  ct_varname = "cellType",
  sample = "batchId",
  ct_sub = NULL,
  iter_max = NULL,
  nu = 1e-04,
  epsilon = NULL,
  truep = NULL,
  weight_basis = TRUE,
  ct_cell_size = NULL,
  Transform_bisque = FALSE,
  grid_search = FALSE,
  search_length = 0.05,
  names_sc_objects = NULL,
  qcthreshold = 0.7,
  verbose = FALSE,
  quality_control = FALSE
)
}
\arguments{
\item{bulk_gene_expression}{Dataframe or matrix of bulk RNA-seq data (genes x individuals)}

\item{single_cell_object}{A matrix or dataframe with the single-cell data. Rows are genes,
columns are samples. Row and column names need to be set. This can also be a list of objects,
if SCDC_ENSEMBLE should be used.}

\item{cell_type_annotations}{A Vector of the cell type annotations. Has to be in the same order
as the samples in single_cell_object. This can also be a list of vectors, if SCDC_ENSEMBLE
should be used.}

\item{batch_ids}{A vector of the ids of the samples or individuals.}

\item{ct_varname}{character string specifying the variable name for 'cell types'.}

\item{sample}{character string specifying the variable name for subject/samples.}

\item{ct_sub}{vector. a subset of cell types that are selected to construct basis matrix. NULL
means that all are used.}

\item{iter_max}{the maximum number of iteration in WNNLS. If the parameter is NULL, the default
value of 1000 for a single single cell object and 2000 for a list is chosen.}

\item{nu}{a small constant to facilitate the calculation of variance}

\item{epsilon}{a small constant number used for convergence criteria. If the parameter is NULL,
the default value of 0.01 for a single single cell object and 0.001 for a list is chosen.}

\item{truep}{true cell-type proportions for bulk samples if known}

\item{weight_basis}{Whether to use the Basis Matrix adjusted for maximal variance weight,
created by the SCDC_basis function}

\item{ct_cell_size}{default is NULL, which means the "library size" is calculated based on the
data. Users can specify a vector of cell size factors corresponding to the ct.sub according to
prior knowledge. The vector should be named: names(ct_cell_size input) should not be NULL.}

\item{Transform_bisque}{The bulk sample transformation from bisqueRNA. Aiming to reduce the
systematic difference between single cells and bulk samples.}

\item{grid_search}{logical. whether to allow grid search method to derive the ENSEMBLE weights.}

\item{search_length}{a number between 0 to 0.5. if using "Grid search", the step length used.
Smaller search.length derives more accurate optimization results.}

\item{names_sc_objects}{A vector with the names of the single cell objects. Only used if
a list of single cell objects is supplied. If it remains NULL, the objects are named by their
index.}

\item{qcthreshold}{The probability threshold used to filter out questionable cells, only used if
quality_control = TRUE.}

\item{verbose}{Whether to create any output.}

\item{quality_control}{Whether to perform the SCDC_qc quality control method.}
}
\value{
Estimated proportion, basis matrix, predicted gene expression levels for bulk samples
}
\description{
This function is to calculate the SCDC deconvolution proportions.
IMPORTANT: No model is needed. Everything is done inside this method.
}
\details{
SCDC_ENSEMBLE can be used by supplying lists to the parameters single_cell_object and
cell_type_annotations. To name the single cell data sets, supply a vector with their
corresponding names to names_sc_objects.

Requires raw read counts.
Works best with multiple cells per single cell patient/subject
}
