% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CibersortX.R
\name{get_signature_matrix_options}
\alias{get_signature_matrix_options}
\title{Creation of the options of the "get signature matrix" docker command}
\usage{
get_signature_matrix_options(
  refsample,
  g_min = 300,
  g_max = 500,
  q_value = 0.01,
  filter = FALSE,
  k_max = 999,
  remake = FALSE,
  replicates = 5,
  sampling = 0.5,
  fraction = 0.75
)
}
\arguments{
\item{refsample}{The filename of the single cell data.}

\item{g_min}{<int> Minimum number of genes per cell type in sig. matrix.}

\item{g_max}{<int> Maximum number of genes per cell type in sig. matrix.}

\item{q_value}{<int> Q-value threshold for differential expression.}

\item{filter}{<bool> Remove non-hematopoietic genes.}

\item{k_max}{<int> Maximum condition number. Will be added to the file name.}

\item{remake}{<bool> Remake signature gene matrix.}

\item{replicates}{<int> Number of replicates to use for building scRNAseq reference file.}

\item{sampling}{<float> Fraction of available single cell GEPs selected using random sampling.}

\item{fraction}{<float> Fraction of cells of same identity showing evidence of expression.}
}
\value{
A string in the correct format for the docker command, containing all parameters.
}
\description{
Creation of the options of the "get signature matrix" docker command
}
