% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CibersortX.R
\name{get_signature_matrix_options}
\alias{get_signature_matrix_options}
\title{Creation of the options of the "get signature matrix" docker command}
\usage{
get_signature_matrix_options(
  refsample,
  G_min = 300,
  G_max = 500,
  q_value = 0.01,
  filter = FALSE,
  k_max = 999,
  remake = FALSE,
  replicates = 5,
  sampling = 0.5,
  fraction = 0.75
)
}
\arguments{
\item{refsample}{The filename of the single cell data}

\item{G_min}{<int> Minimum number of genes per cell type in sig. matrix (default: 300)}

\item{G_max}{<int> Maximum number of genes per cell type in sig. matrix (default: 500)}

\item{q_value}{<int> Q-value threshold for differential expression (default: 0.01)}

\item{filter}{<bool> Remove non-hematopoietic genes (default: FALSE)}

\item{k_max}{<int> Maximum condition number (default: 999). Will be added to the file name.}

\item{remake}{<bool> Remake signature gene matrix (default: False)}

\item{replicates}{<int> Number of replicates to use for building scRNAseq reference file
(default: 5)}

\item{sampling}{<float> Fraction of available single cell GEPs selected using random sampling
(default: 0.5)}

\item{fraction}{<float> Fraction of cells of same identity showing evidence of expression
(default: 0.75)}
}
\value{
A string in the correct format for the docker command, containing all parameters.
}
\description{
Creation of the options of the "get signature matrix" docker command
}
