% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{install_scaden}
\alias{install_scaden}
\title{Install Scaden}
\usage{
install_scaden()
}
\arguments{
\item{conda}{}
}
\value{

}
\description{
Install Scaden into a previously defined python environment.
A python environment can be defined by set_virtualenv() or set_python() command.
Alternatively a new environment can be created via create_virtualenv() method.
}
