% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{makeBenchmarkingScatterplot}
\alias{makeBenchmarkingScatterplot}
\title{Make a Scatterplot for Benchmarking}
\usage{
makeBenchmarkingScatterplot(result_list, ref_data, file_name = NULL)
}
\arguments{
\item{result_list}{named list containing all deconvolution results that should be considered,
cell type annotations need to contain the same cell types as the ones in ref_data}

\item{ref_data}{reference cell types which are used as the ground truth}

\item{file_name}{(optional) plot is saved in this file}
}
\value{
the ggplot object
}
\description{
Make a Scatterplot for Benchmarking
}
\examples{
data("single_cell_data_1")
data("cell_type_annotations_1")
data("batch_ids_1")
data("bulk")
data("RefData")

common_genes <- intersect(rownames(single_cell_data_1),rownames(bulk))[1:2000]

single_cell_data <- single_cell_data_1[common_genes,1:500]
cell_type_annotations <- cell_type_annotations_1[1:500]
batch_ids <- batch_ids_1[1:500]
bulk <- bulk[common_genes,]

names(RefData) <- c("T", "Mono", "B", "DC", "NK")
RefData <- RefData[,order(colnames(RefData))]

sig_bisque <- build_model(
  single_cell_data, cell_type_annotations, "bisque",
  batch_ids
)
res_bisque <- deconvolute(
  bulk, sig_bisque, "bisque", single_cell_data,
  cell_type_annotations, batch_ids
)
#Merging the two T cell props
res_bisque <- cbind(res_bisque,T = (res_bisque[,"CD4 T"]+res_bisque[,"CD8 T"]))[,-c(2,3)]

res_scdc <- deconvolute(bulk, NULL, "scdc", batch_ids,
                         single_cell_object = single_cell_data,
                         cell_type_annotations = cell_type_annotations
)
#Merging the two T cell props
res_scdc <- cbind(res_scdc,T = (res_scdc[,"CD4 T"]+res_scdc[,"CD8 T"]))[,-c(2,3)]

result_list <- list(SCDC = res_scdc, Bisque = res_bisque)
makeBenchmarkingScatterplot(result_list, RefData)
# Alternative if you want to save the plot in a file
# makeBenchmarkingScatterplot(result_list, RefData, "predictionVsGroundtruth.png")
}
