% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{makeBenchmarkingScatterplot}
\alias{makeBenchmarkingScatterplot}
\title{Make a Scatterplot for Benchmarking}
\usage{
makeBenchmarkingScatterplot(result_list, file_name)
}
\arguments{
\item{result_list}{named list containing all deconvolution results that should be considered,
cell type annotations need to contain "T cell", "DC", "Mono", "NK", "B cell" in any way to be included}

\item{file_name}{(optional) plot is saved in this file}
}
\value{
the ggplot object
}
\description{
Make a Scatterplot for Benchmarking
}
