% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotDeconvResult}
\alias{plotDeconvResult}
\title{Plot deconvolution results as a barplot}
\usage{
plotDeconvResult(deconv_result, method_name = "", file_name = NULL)
}
\arguments{
\item{deconv_result}{result from deconvolution()}

\item{method_name}{(optional) title of plot is set to the name of the method used for
deconvolution}

\item{file_name}{(optional) plot is saved in this file}
}
\value{
the ggplot object
}
\description{
Plot deconvolution results as a barplot
}
\examples{
data("single_cell_data_1")
data("cell_type_annotations_1")
data("batch_ids_1")
data("bulk")
data("RefData")

common_genes <- intersect(rownames(single_cell_data_1),rownames(bulk))[1:2000]

single_cell_data <- single_cell_data_1[common_genes,1:500]
cell_type_annotations <- cell_type_annotations_1[1:500]
batch_ids <- batch_ids_1[1:500]
bulk <- bulk[common_genes,]

model <- build_model(single_cell_data, cell_type_annotations, "bisque", batch_ids)
deconvolution <- deconvolute(
  bulk, model, "bisque", single_cell_data,
  cell_type_annotations, batch_ids
)
plotDeconvResult(deconvolution, "Bisque")
}
