% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{scaden_build_model}
\alias{scaden_build_model}
\title{Builds Scaden Model from scRNA data}
\usage{
scaden_build_model(
  count_data,
  celltype_labels,
  gene_labels,
  bulk_data = NULL,
  model_path = NULL,
  verbose = F,
  ...
)
}
\arguments{
\item{count_data}{Matrix or dataframe of scRNA data. Rows=cells and columns=genes}

\item{celltype_labels}{Vector of celltype labels. Order corresponds to rows in count_data matrix.}

\item{gene_labels}{Vector of gene labels. Order corresponds to columns in count_data matrix.}

\item{bulk_data}{Matrix of bulk RNA expression. (genes x samples)}

\item{model_path}{Path where model directory should be created.}
}
\value{

}
\description{
The model is saved in a defined directory.
}
