% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{scaden_predict}
\alias{scaden_predict}
\title{Predict cell proportions}
\usage{
scaden_predict(model_dir, bulk_data, verbose = FALSE)
}
\arguments{
\item{model_dir}{Directory where model is saved}

\item{bulk_data}{Bulk RNA-seq data. (genes x bulk_RNA samples)}

\item{verbose}{Defines verbosity of function call (default: false)}
}
\value{
Cell type fractions per sample
}
\description{
Predicts cell proportions in bulk RNA sample.
}
