% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{scaden_simulate}
\alias{scaden_simulate}
\title{Simulates training data from scRNA data}
\usage{
scaden_simulate(
  celltype_labels,
  gene_labels,
  single_cell_object,
  cells = 100,
  samples = 1000,
  dataset_name = "scaden",
  verbose = F
)
}
\arguments{
\item{celltype_labels}{Vector of celltype labels. Order corresponds to rows in single_cell_object matrix.}

\item{gene_labels}{Vector of gene labels. Order corresponds to columns in single_cell_object matrix.}

\item{single_cell_object}{Matrix or dataframe of scRNA data. Rows=cells and columns=genes}
}
\value{

}
\description{
Simulates training data from scRNA data
}
