% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaden.R
\name{scaden_train}
\alias{scaden_train}
\title{Trains a Scaden model}
\usage{
scaden_train(
  h5ad_processed,
  temp_dir = NULL,
  batch_size = 128,
  learning_rate = 1e-04,
  model_path = NULL,
  steps = 5000,
  verbose = FALSE
)
}
\arguments{
\item{h5ad_processed}{pre-processed training data}

\item{temp_dir}{The temporary directory to use for the computations (optional)}

\item{batch_size}{Batch size to use for training. Default: 128}

\item{learning_rate}{Learning rate used for training. Default: 0.0001}

\item{model_path}{Path to store the model in}

\item{steps}{Number of training steps}

\item{verbose}{Whether to produce an output on the console. (default: false)}
}
\value{
Scaden model
}
\description{
Trains a Scaden model on pre-processed training data.
}
