% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper.R
\name{set_cibersortx_credentials}
\alias{set_cibersortx_credentials}
\title{Setting the CibersortX Credentials}
\usage{
set_cibersortx_credentials(email, token)
}
\arguments{
\item{email}{The email which was used to register on the website.}

\item{token}{The token provided by the website.}
}
\description{
Setting the CibersortX Credentials
}
