% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{set_python}
\alias{set_python}
\title{Defines python path.}
\usage{
set_python(path_to_python_binaries)
}
\arguments{
\item{path_to_python_binaries}{Path to python binaries}
}
\description{
Python3 is needed for scanpy and scaden to work.
}
