% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{singlecellexperiment_to_anndata}
\alias{singlecellexperiment_to_anndata}
\title{Convert SingleCellExperiment to AnnData object}
\usage{
singlecellexperiment_to_anndata(sce, X_name = NULL)
}
\arguments{
\item{sce}{object of type SingleCellExperiment}

\item{X_name}{Name of the assay in the SingleCellExperiment to interpret as the X matrix of the AnnData object (default: first in assay list)}
}
\value{
AnnData object
}
\description{
Convert SingleCellExperiment to AnnData object
}
