% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{singlecellexperiment_to_matrix}
\alias{singlecellexperiment_to_matrix}
\title{Convert SingleCellExperiment to Matrix + annotation vector for celltypes}
\usage{
singlecellexperiment_to_matrix(sce, assay_name = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{assay_name}{name of the assay that should be returned as matrix. (default: first in assay list)}
}
\value{
named list: ..$matrix: matrix object, ..$annotation_vector
}
\description{
Convert SingleCellExperiment to Matrix + annotation vector for celltypes
}
