% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnData.R
\name{write_anndata}
\alias{write_anndata}
\title{Write anndata object (.h5ad format)}
\usage{
write_anndata(data, path)
}
\arguments{
\item{data}{AnnData object}

\item{path}{path where AnnData object should be written to (.h5ad format)}
}
\description{
Write anndata object (.h5ad format)
}
